chore: add quantel tests
